+++ examples/polybench/gemm/src/lib.rs
@@ -5,21 +5,21 @@ pub fn gemm_kernel(
     a: &[f32],
     b: &[f32],
     c: &mut [f32],
-    ni: usize,
-    nj: usize,
-    nk: usize,
+    ni: u32,
+    nj: u32,
+    nk: u32,
     alpha: f32,
     beta: f32,
 ) {
-    let mut c = chunk_mut(c, Map2D::new(nj));
-    let j = (block_id::<DimX>() * block_dim::<DimX>() + thread_id::<DimX>()) as usize;
-    let i = (block_id::<DimY>() * block_dim::<DimY>() + thread_id::<DimY>()) as usize;
+    let mut c = chunk_mut(c, Map2D::new(nj as usize));
+    let j = block_id::<DimX>() * block_dim::<DimX>() + thread_id::<DimX>();
+    let i = block_id::<DimY>() * block_dim::<DimY>() + thread_id::<DimY>();
 
     if i < ni && j < nj {
         let mut val = c[(0, 0)] * beta;
-        let mut k: usize = 0;
+        let mut k: u32 = 0;
         while k < nk {
-            val += alpha * a[i * nk + k] * b[k * nj + j];
+            val += alpha * a[(i * nk + k) as usize] * b[(k * nj + j) as usize];
             k += 1;
         }
         c[(0, 0)] = val;
@@ -60,7 +60,7 @@ mod tests {
             let config =
                 gpu_host::gpu_config!(grid_x, grid_y, 1, block_size, block_size, 1, 0);
             gemm_kernel::launch(
-                config, ctx, m, &d_a, &d_b, &mut d_c, ni, nj, nk, alpha, beta,
+                config, ctx, m, &d_a, &d_b, &mut d_c, ni as u32, nj as u32, nk as u32, alpha, beta,
             )
             .expect("kernel launch failed");
